+++ internal/features/tests/tests_feature.go
@@ -66,10 +66,13 @@ func (f *TestsFeature) Start(ctx context.Context) {
 	didChangeDone := make(chan job.IDs, 10)
 	didChangeWatchedDone := make(chan job.IDs, 10)
 
+	providerSchemaChangeDone := make(chan job.IDs, 10)
+
 	discover := f.bus.OnDiscover(topic, nil)
 	didOpen := f.bus.OnDidOpen(topic, didOpenDone)
 	didChange := f.bus.OnDidChange(topic, didChangeDone)
 	didChangeWatched := f.bus.OnDidChangeWatched(topic, didChangeWatchedDone)
+	providerSchemaChange := f.bus.OnProviderSchemaChange(topic, providerSchemaChangeDone)
 
 	go func() {
 		for {
@@ -89,6 +92,10 @@ func (f *TestsFeature) Start(ctx context.Context) {
 				// TODO? collect errors
 				spawnedIds, _ := f.didChangeWatched(didChangeWatched.Context, didChangeWatched.RawPath, didChangeWatched.ChangeType, didChangeWatched.IsDir)
 				didChangeWatchedDone <- spawnedIds
+			case providerSchemaChange := <-providerSchemaChange:
+				// TODO? collect errors
+				spawnedIds, _ := f.providerSchemaChange(providerSchemaChange.Context, providerSchemaChange.Dir)
+				providerSchemaChangeDone <- spawnedIds
 
 			case <-ctx.Done():
 				return